refactoring:change getRecommendations from UserService to FilmService, stay method for single SQL query
