+++ src/main/java/ru/yandex/practicum/filmorate/controllers/UserController.java
@@ -7,8 +7,8 @@ import org.springframework.web.bind.annotation.*;
 import ru.yandex.practicum.filmorate.model.Film;
 import ru.yandex.practicum.filmorate.model.MessageResponse;
 import ru.yandex.practicum.filmorate.model.User;
+import ru.yandex.practicum.filmorate.services.FilmService;
 import ru.yandex.practicum.filmorate.services.UserService;
-
 import javax.validation.Valid;
 import java.util.Collection;
 
@@ -18,6 +18,7 @@ import java.util.Collection;
 @RequestMapping("/users")
 public class UserController {
     private final UserService userService;
+    private final FilmService filmService;
 
     @GetMapping
     public Collection<User> findAllUsers() {
@@ -36,7 +37,7 @@ public class UserController {
     @GetMapping ("/{id}/recommendations")
     public Collection<Film> getRecommendations(@PathVariable Long id) {
         log.info("GET /{}/recommendations - запрос рекомендаций для пользователя.", id);
-        return userService.getRecommendations(id);
+        return filmService.getRecommendations(id);
     }
 
     @PostMapping
