+++ src/main/java/ru/yandex/practicum/filmorate/storage/film/FilmStorage.java
@@ -2,6 +2,7 @@ package ru.yandex.practicum.filmorate.storage.film;
 
 import ru.yandex.practicum.filmorate.model.Film;
 
+import java.util.Collection;
 import java.util.List;
 import java.util.Optional;
 
@@ -18,4 +19,6 @@ public interface FilmStorage {
     Film updateFilm(Film film);
 
     Boolean checkFilm(Long filmId);
+
+    Collection<Film> getRecommendations(Long userId);
 }
