+++ src/main/java/ru/yandex/practicum/filmorate/storage/user/UserDbStorage.java
@@ -120,88 +120,6 @@ public class UserDbStorage implements UserStorage {
     }
 
     @Override
-    public Collection<Film> getRecommendations(Long userId) {
-        //uf - user films table
-        String sql = "SELECT f.*, m.mpa_name, COALESCE(s.count_like, 0) AS rate " +
-                "FROM films AS f " +
-                "LEFT JOIN mpas AS m on m.mpa_id = f.mpa_id " +
-                "LEFT JOIN (SELECT fl.film_id, " +
-                "COUNT(fl.user_id) AS count_like " +
-                "FROM film_like AS fl " +
-                "GROUP BY fl.film_id) AS s ON f.film_id=s.film_id " +
-                "INNER JOIN " +
-                    "(SELECT film_id " +
-                    "FROM film_like " +
-                    "WHERE user_id = " +
-                        "(SELECT user_id " +
-                        "FROM film_like WHERE film_id IN " +
-                            "(SELECT film_id FROM film_like WHERE user_id = ?) " +
-                        "AND user_id <> ? " +
-                        "GROUP BY user_id " +
-                        "ORDER BY COUNT(user_id) DESC " +
-                        "LIMIT 1) " +
-                    "AND film_id NOT IN " +
-                        "(SELECT film_id FROM film_like WHERE user_id = ?)) " +
-                "AS uf ON uf.film_id = f.film_id "+
-                "ORDER BY rate DESC";
-        Collection<Film> films = jdbcTemplate.query(sql, filmDbStorage::makeFilm, userId, userId, userId);
-        return films;
-    }
-
-    @Override
-    public Collection<Film> getRecommendationsWithSeparateMethods(Long userId) {
-        Collection<Film> recommendedFilms = new ArrayList<>();
-        if (!isUserSetLike(userId)) return recommendedFilms;
-        Long anotherUser = findRecommendedUser(userId);
-        if (anotherUser == null) return recommendedFilms;
-        Collection<Film> anotherUserFilms = findUserFilms(anotherUser);
-        Collection<Film> userFilms = findUserFilms(userId);
-        for (Film film : anotherUserFilms) {
-            if (!userFilms.contains(film)) {
-                recommendedFilms.add(film);
-            }
-        }
-        return recommendedFilms;
-    }
-
-    private Long findRecommendedUser(Long userId) {
-        String findAnotherUser = "SELECT user_id FROM film_like WHERE film_id IN " +
-                "(SELECT film_id FROM film_like WHERE user_id = ?) " +
-                "AND user_id <> ? " +
-                "GROUP BY user_id " +
-                "ORDER BY COUNT(user_id) DESC " +
-                "LIMIT 1";
-        try {
-            return jdbcTemplate.queryForObject(findAnotherUser, Long.class, userId, userId);
-        } catch (Exception e) {
-            return null;
-        }
-    }
-
-    private Boolean isUserSetLike(Long userId) {
-        String findLike = "SELECT EXISTS(SELECT 1 FROM film_like WHERE user_id=?)";
-        return jdbcTemplate.queryForObject(findLike, (rs, rowNum) -> rs.getBoolean(1), userId);
-    }
-
-    private Collection<Film> findUserFilms(Long userId) {
-        String sql = "SELECT f.*, m.mpa_name, COALESCE(s.count_like, 0) AS rate " +
-                "FROM films AS f " +
-                "LEFT JOIN mpas AS m on m.mpa_id = f.mpa_id " +
-                "LEFT JOIN (SELECT fl.film_id, " +
-                "COUNT(fl.user_id) AS count_like " +
-                "FROM film_like AS fl " +
-                "GROUP BY fl.film_id) AS s ON f.film_id=s.film_id " +
-                "INNER JOIN " +
-                    "(SELECT film_id " +
-                    "FROM film_like " +
-                    "WHERE user_id = ?)" +
-                    "AS uf ON uf.film_id = f.film_id "+
-                "ORDER BY rate DESC";
-        Collection<Film> films = jdbcTemplate.query(sql, filmDbStorage::makeFilm, userId);
-        return films;
-    }
-
-    @Override
     public Boolean checkUser(Long userId) {
         String sql = "SELECT EXISTS(SELECT 1 FROM users WHERE user_id=?)";
         return jdbcTemplate.queryForObject(sql, (rs, rowNum) -> rs.getBoolean(1), userId);
