+++ src/main/java/ru/yandex/practicum/filmorate/storage/user/UserStorage.java
@@ -25,7 +25,4 @@ public interface UserStorage {
 
     Boolean checkUser(Long userId);
 
-    Collection<Film> getRecommendations(Long userId);
-
-    Collection<Film> getRecommendationsWithSeparateMethods(Long userId);
 }
